+++ src/pages/MyTeam/components/MyTeamList/MyTeamCard/MyTeamDeletionButton.tsx
@@ -11,7 +11,6 @@ export default function MyTeamDeletionButton({ teamId }: { teamId: string }) {
         deletion.mutate(teamId);
       }}
     >
-      {/**버튼 내용물은 바꾸는걸로... 단체로 이러니까 좀 짜치는 너낌... */}
       <RedHeart />
     </button>
   );
